Refactoring to support PlayerGroups

+++ client/src/Pixi/handlers/incoming/IncomingBoardUpdateMessages.ts
@@ -15,7 +15,12 @@ const IncomingBoardUpdateMessages: { [index in BoardUpdateMessageType]: Incoming
 			return
 		}
 
-		const card = new RenderedUnit(RenderedCard.fromMessage(data.card), Core.getPlayer(data.ownerId))
+		const ownerGroup = Core.getPlayerGroupOrNull(data.ownerId)
+		if (!ownerGroup) {
+			throw new Error(`No player group with id ${data.ownerId}`)
+		}
+
+		const card = new RenderedUnit(RenderedCard.fromMessage(data.card), ownerGroup)
 		Core.input.destroyLimboCard(data.card)
 		Core.board.insertUnit(card, data.rowIndex, data.unitIndex)
 	},
@@ -40,7 +45,7 @@ const IncomingBoardUpdateMessages: { [index in BoardUpdateMessageType]: Incoming
 	},
 
 	[BoardUpdateMessageType.ROW_OWNER]: (data: BoardRowMessage) => {
-		Core.board.rows[data.index].owner = Core.getPlayerOrNull(data.ownerId)
+		Core.board.rows[data.index].owner = Core.getPlayerGroupOrNull(data.ownerId)
 	},
 
 	[BoardUpdateMessageType.ROW_BUFF_ADD]: (data: OpenRowBuffMessage) => {
